optimize db package
