+++ database/cli/src/main.rs
@@ -15,13 +15,9 @@ use reqwest::StatusCode;
 use std::io::Write;
 use std::time::Duration;
 
-// TODO move to db package
-// now if imported from there, a lot of unneeded stuff gets compiled
-pub const MANAGER_ADDR: &str = "localhost:3000";
-
 fn main() -> anyhow::Result<()> {
     let client = reqwest::blocking::Client::new();
-    let addr = MANAGER_ADDR.to_string();
+    let addr = db::MANAGER_ADDR.to_string();
 
     let server = Server { client, addr };
